# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/admin/user/active": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API"
                ],
                "summary": "用户激活账户，设置密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户新密码",
                        "name": "new_password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "AdminUser",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.AdminUserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/admin/user/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API"
                ],
                "summary": "创建后台账户",
                "parameters": [
                    {
                        "description": "用户名",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "邮箱",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "权限",
                        "name": "auth",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "AdminUser",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.AdminUserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/admin/user/reset_password": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API"
                ],
                "summary": "重置用户密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "AdminUser",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.AdminUserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/admin/user/search": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API"
                ],
                "summary": "查询用户",
                "parameters": [
                    {
                        "description": "用户名称",
                        "name": "name",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "用户email",
                        "name": "email",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "用户状态",
                        "name": "status",
                        "in": "body",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "每页数量",
                        "name": "size",
                        "in": "body",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "当前页数，从1开始",
                        "name": "page",
                        "in": "body",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search Result",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/admin.AdminUserResult"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/admin/user/update_base_info": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API"
                ],
                "summary": "更新用户信息",
                "parameters": [
                    {
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "邮箱",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "权限",
                        "name": "auth",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "admin user",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.AdminUserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/admin/user/update_status": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API"
                ],
                "summary": "修改用户状态",
                "parameters": [
                    {
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "用户状态",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "admin user",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.AdminUserResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "admin.AdminUserAuth": {
            "type": "integer",
            "enum": [
                0,
                -1
            ],
            "x-enum-comments": {
                "AdminAuth_None": "某些api不限制权限，只要求登录的时候，使用这个权限"
            },
            "x-enum-varnames": [
                "AdminAuth_None",
                "AdminAuth_Admin"
            ]
        },
...
        "admin.AdminUserResult": {
            "type": "object",
            "properties": {
                "authorization": {
                    "$ref": "#/definitions/admin.AdminUserAuth"
                },
                "create_at": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/admin.AdminUserStatus"
                }
            }
        },
        "admin.AdminUserStatus": {
            "type": "integer",
            "enum": [
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "AdminStatus_Normal",
                "AdminStatus_ResetPassword",
                "AdminStatus_Ban"
            ]
        },
        "api.AdminUserResponse": {
            "type": "object",
            "required": [
                "user_id"
            ],
            "properties": {
                "auth": {
                    "$ref": "#/definitions/admin.AdminUserAuth"
                },
                "email": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/admin.AdminUserStatus"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "api.JSONResult": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/api.ResultCode"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "api.ResultCode": {
            "type": "integer",
            "enum": [
                0,
                -1
            ],
            "x-enum-varnames": [
                "Success",
                "Error"
            ]
        },
...
    }
}